Fixed 32-bit bug in ELF Rela sections
git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@2539 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ src/frontend/ExecFormats/ROSE_ExecELF.C
@@ -1699,7 +1699,7 @@ SgAsmElfRelaEntry::encode(ByteOrder sex, Elf64RelaEntry_disk *disk)
 {
     host_to_disk(sex, p_r_offset, &(disk->r_offset));
     host_to_disk(sex, p_r_addend, &(disk->r_addend));
-    uint64_t info = (p_sym<<32) | (p_type & 0xffffffff);
+    uint64_t info = ((uint64_t)p_sym<<32) | (p_type & 0xffffffff);
     host_to_disk(sex, info, &(disk->r_info));
     return disk;
 }
